requirements.txt
streamlit
mysql-connector-python
pandas
plotly


# Copy and paste this in your terminal: pip install -r requirements.txt